requirements.txt
# Django
Django==4.2.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# API
djangorestframework==3.14.0

# CORS
django-cors-headers==4.3.1

# For images/media (if needed)
Pillow==10.1.0